+++ src/main/java/com/example/commerce/pessoa/PessoaService.java
@@ -4,14 +4,13 @@ import java.util.List;
 
 public interface PessoaService {
 
-    Pessoa findOne(Long id);
+    Pessoa findOne(Integer id);
 
     List<Pessoa> findAll();
 
     Pessoa save (Pessoa pessoa);
 
-
-    void delete(Long id);
+    void delete(Integer id);
 
     List<Pessoa> complete(String query, String categoria);
 }
